+++ r-naming.tf
@@ -1,0 +1,41 @@
+resource "azurecaf_name" "vm" {
+  name          = var.stack
+  resource_type = "azurerm_windows_virtual_machine"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "vm"])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}
+
+resource "azurecaf_name" "pub_ip" {
+  name          = var.stack
+  resource_type = "azurerm_public_ip"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "pubip"])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}
+
+resource "azurecaf_name" "nic" {
+  name          = var.stack
+  resource_type = "azurerm_network_interface"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "nic"])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}
+
+resource "azurecaf_name" "disk" {
+  for_each = var.storage_data_disk_config
+
+  name          = var.stack
+  resource_type = "azurerm_managed_disk"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, each.key])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}