+++ versions.tf
@@ -1,5 +1,5 @@
 terraform {
-  required_version = "> 0.12.26"
+  required_version = ">= 0.13"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
@@ -8,6 +8,10 @@ terraform {
     null = {
       source  = "hashicorp/null"
       version = "~> 3.0"
+    }
+    azurecaf = {
+      source  = "aztfmod/azurecaf"
+      version = "~> 1.1"
     }
   }
 }